+++ pkg/applicationserver/io/packages/loradms/v1/package.go
@@ -17,7 +17,6 @@ package loraclouddevicemanagementv1
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"time"
 
 	"github.com/gogo/protobuf/types"
@@ -39,14 +38,9 @@ import (
 const PackageName = "lora-cloud-device-management-v1"
 
 // DeviceManagementPackage is the LoRa Cloud Device Management application package.
-//
-// TODO: Once https://github.com/TheThingsNetwork/lorawan-stack/issues/2533 is implemented,
-// change io.Server usage to an interface that extends io.Server, but with the HTTPClient()
-// method added, and remove the custom client.
 type DeviceManagementPackage struct {
 	server   io.Server
 	registry packages.Registry
-	client   *http.Client
 }
 
 // RegisterServices implements packages.ApplicationPackageHandler.
@@ -121,7 +115,12 @@ func (p *DeviceManagementPackage) sendUplink(ctx context.Context, up *ttnpb.Appl
 	logger := log.FromContext(ctx)
 	eui := objects.EUI(*up.DevEUI)
 
-	client, err := api.New(p.client, api.WithToken(data.token), api.WithBaseURL(data.serverURL))
+	httpClient, err := p.server.HTTPClient(ctx)
+	if err != nil {
+		logger.WithError(err).Debug("Failed to create HTTP client")
+		return err
+	}
+	client, err := api.New(httpClient, api.WithToken(data.token), api.WithBaseURL(data.serverURL))
 	if err != nil {
 		logger.WithError(err).Debug("Failed to create API client")
 		return err
@@ -312,9 +311,6 @@ func New(server io.Server, registry packages.Registry) packages.ApplicationPacka
 	return &DeviceManagementPackage{
 		server:   server,
 		registry: registry,
-		client: &http.Client{
-			Timeout: 30 * time.Second,
-		},
 	}
 }