+++ pkg/frequencyplans/frequencyplans_test.go
@@ -16,6 +16,7 @@ package frequencyplans_test
 
 import (
 	"fmt"
+	"net/http"
 	"testing"
 	"time"
 
@@ -34,7 +35,7 @@ func boolPtr(v bool) *bool                       { return &v }
 func durationPtr(v time.Duration) *time.Duration { return &v }
 
 func Example() {
-	fetcher, err := fetch.FromHTTP("https://raw.githubusercontent.com/TheThingsNetwork/lorawan-frequency-plans", true)
+	fetcher, err := fetch.FromHTTP(http.DefaultTransport, "https://raw.githubusercontent.com/TheThingsNetwork/lorawan-frequency-plans", true)
 	if err != nil {
 		panic(err)
 	}