+++ pkg/networkserver/networkserver.go
@@ -191,6 +191,13 @@ func New(c *component.Component, conf *Config, opts ...Option) (*NetworkServer, 
 			return c.GetTLSClientConfig(ctx)
 		}
 		interopConf.BlobConfig = c.GetBaseConfig(ctx).Blob
+		if interopConf.Transport == nil {
+			tr, err := c.HTTPTransport(ctx)
+			if err != nil {
+				return nil, err
+			}
+			interopConf.Transport = tr
+		}
 
 		interopCl, err = interop.NewClient(ctx, interopConf)
 		if err != nil {