+++ pkg/web/oauthclient/oauthclient.go
@@ -17,7 +17,6 @@ package oauthclient
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"net/url"
 	"strings"
 
@@ -56,16 +55,12 @@ func (oc *OAuthClient) getMountPath() string {
 	return path
 }
 
-func (oc *OAuthClient) withTLSClientConfig(ctx context.Context) (context.Context, error) {
-	config, err := oc.component.GetTLSClientConfig(ctx)
+func (oc *OAuthClient) withHTTPClient(ctx context.Context) (context.Context, error) {
+	client, err := oc.component.HTTPClient(ctx)
 	if err != nil {
 		return nil, err
 	}
-	return context.WithValue(ctx, oauth2.HTTPClient, &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: config,
-		},
-	}), nil
+	return context.WithValue(ctx, oauth2.HTTPClient, client), nil
 }
 
 // New returns a new OAuth client instance.